refactor(server): extract export-format builders to export_builders.py (router-split, round-5 #51 / R5-25 PR10)
Third leaf module of the APIRouter split (#51), stacked on webguard.py (PR9).
Moves the pure export serialisers that turn a media record into CSV / EDL /
FCPXML / SRT / VTT text — plus the timecode + framerate math they share — into a
state-free leaf module the export routers and server both import, so the coming
router peel doesn't hit the router→server→router cycle:

  CSV      _CSV_FORMULA_PREFIXES / _csv_safe / _parse_frame_tags /
           _build_metadata_csv
  EDL      _edl_reel / _media_streams / _edl_comment / _edl_timecode /
           _start_tc_seconds / _edl_fps_warning
  SRT/VTT  _subtitle_ts / _subtitle_text
  FCPXML   _fcpxml_rational

Every function is pure except _build_metadata_csv, which reads the DB via `db` —
itself a leaf (db.py does not import server), so no cycle. export_builders
depends only on json + db + stdlib.

Deliberately LEFT in server.py: _attachment_headers (Content-Disposition builder,
an HTTP-response concern) and _log_safe (terminal-log sanitiser) — neither is an
export-format serialiser; they'll travel with the export router later.

server.py re-exports all 13 names so existing call sites / tests referencing
server._build_metadata_csv etc. keep working unchanged. Net server.py -251 lines.
New test module pins the leaf boundary, identity re-export, the HTTP/log helpers
staying put, and a representative slice of the injection-hardening + timecode
format behaviour (CSV formula defusing, EDL reel/comment control-char stripping,
subtitle cue-boundary neutralising, DF/NDF timecode, NTSC rational, mixed-fps
warning, frame_tags JSON+legacy parsing). 825 passed.
